class names are refactored and unit test added for helper classes
