+++ app/src/main/java/com/ss/stepperview/sample/data/StudyTaskExample.kt
@@ -5,23 +5,20 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Shapes
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.HorizontalAlignmentLine
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.rememberImagePainter
-import com.ss.stepperview.Step
 import com.ss.stepperview.StepAlignment
-import com.ss.stepperview.StepperView
-import com.ss.stepperview.StepsPerRow
+import com.ss.stepperview.view.StepperView
+import com.ss.stepperview.view.StepsPerRow
 import com.ss.stepperview.sample.R
 import com.ss.stepperview.sample.ui.theme.green01
 import com.ss.stepperview.sample.ui.theme.purple01
@@ -96,15 +93,15 @@ fun DefaultPreview() {
 @Composable
 fun StudyTaskStepperView(studyTaskList: ArrayList<StudyTask>){
     StepperView(items = studyTaskList,
-    stepsPerRow = StepsPerRow.ONE) {
+    stepsPerRow = StepsPerRow.TWO) {
         studyTaskList.forEachIndexed{ index, it ->
-            var alignment = StepAlignment.LEFT
+            var alignment = StepAlignment.RIGHT
             if( index % 2 == 0 || index % 5 == 0 ){
                 alignment = StepAlignment.RIGHT
             }
-//            StudyTaskTimeUI(studyTask = it,
-//                modifier = Modifier
-//                    .align(StepAlignment.RIGHT))
+            StudyTaskTimeUI(studyTask = it,
+                modifier = Modifier
+                    .align(StepAlignment.LEFT))
             StudyTaskUI(selected = false, studyTask = it,
                 modifier = Modifier
                     .align(alignment))
