--- stepperview/src/main/java/com/ss/stepperview/StepperView.kt
@@ -1,590 +1,0 @@
-package com.ss.stepperview
-
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.ButtonDefaults
-import androidx.compose.material.OutlinedButton
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.*
-import androidx.compose.ui.unit.Constraints
-import androidx.compose.ui.unit.Density
-import androidx.compose.ui.unit.dp
-import com.ss.stepperview.StepIndicatorScopeInstance.align
-
-private const val LAYOUTID_STEPPERVIEW_INDICATOR = "LAYOUTID_STEPPERVIEW_INDICATOR"
-private const val LAYOUTID_STEPPERVIEW_LINE = "LAYOUTID_STEPPERVIEW_LINE"
-private const val VERTICAL_SPACING = 20
-
-enum class StepsPerRow(val noOfSteps: Int) {
-    ONE(1), TWO(2)
-}
-
-@Composable
-fun StepperView(
-    items: List<Any>,
-    stepsPerRow: StepsPerRow = StepsPerRow.ONE,
-    content: @Composable StepScope.() -> Unit
-) {
-    StepperViewLayout(
-        Modifier,
-        stepsPerRow = stepsPerRow,
-        indicatorContent = {
-            repeat(items.size) {
-                StepperViewIndicator(
-                    modifier = Modifier
-                        .layoutId(
-                            LAYOUTID_STEPPERVIEW_INDICATOR.plus(it)
-                        )
-                        .align(StepIndicatorAlignment.TOP)
-                )
-            }
-        },
-        lineContent = {
-            repeat(items.size - 1) {
-                StepperViewLine(modifier = Modifier.layoutId(LAYOUTID_STEPPERVIEW_LINE.plus(it)))
-            }
-        },
-        stepContent = {
-            StepScopeInstance.content()
-        }
-    )
-
-}
-
-@Composable
-fun Step(
-    content: @Composable () -> Unit
-) {
-    content()
-}
-
-private val Measurable.stepAlignment
-    get() = (parentData as? StepAlignmentData)?.stepAlignment ?: StepAlignment.RIGHT
-
-private val Measurable.stepIndicatorAlignment
-    get() = (parentData as? StepIndicatorAlignmentData)?.stepIndicatorAlignment
-        ?: StepIndicatorAlignment.CENTER
-
-
-@Composable
-fun StepperViewLayout(
-    modifier: Modifier = Modifier,
-    stepsPerRow: StepsPerRow,
-    indicatorContent: @Composable () -> Unit,
-    lineContent: @Composable () -> Unit,
-    stepContent: @Composable StepScope.() -> Unit
-) {
-    Layout(
-        content = {
-            indicatorContent()
-            lineContent()
-            StepScopeInstance.stepContent()
-        },
-        modifier = modifier
-    ) { measurables, constraints ->
-
-        val stepsMeasurables = measurables
-            .filter {
-                it.layoutId.toString().contains(LAYOUTID_STEPPERVIEW_LINE)
-                    .not() && it.layoutId.toString().contains(LAYOUTID_STEPPERVIEW_INDICATOR)
-                    .not()
-            }
-        val indicatorMeasurables = measurables
-            .filter { it.layoutId.toString().contains(LAYOUTID_STEPPERVIEW_INDICATOR) }
-        val lineMeasurables = measurables
-            .filter { it.layoutId.toString().contains(LAYOUTID_STEPPERVIEW_LINE) }
-
-
-        layout(constraints.maxWidth, constraints.maxHeight) {
-            val stepRowsMeasureAndPlaceHelpersImpl = StepRowMeasureAndPlaceHelpersImpl()
-            val stepRows = Rows(stepsMeasurables, stepsPerRow, stepRowsMeasureAndPlaceHelpersImpl)
-            val indicatorMeasureAndPlaceHelpersImpl = IndicatorMeasureAndPlaceHelpersImpl(stepRows)
-            val indicators =
-                StepIndicators(indicatorMeasurables, indicatorMeasureAndPlaceHelpersImpl)
-            val lineMeasureAndPlaceHelpersImpl = LineMeasureAndPlaceHelpersImpl(indicators)
-            val lines = StepLines(lineMeasurables, lineMeasureAndPlaceHelpersImpl)
-
-            stepRows.measure(
-                constraints.copy(
-                    maxWidth = constraints.maxWidth - Math.max(
-                        indicators.maxIntrinsicWidth,
-                        lines.maxIntrinsicWidth
-                    )
-                )
-            )
-            indicators.measure(constraints)
-
-            stepRows.place.invoke(
-                this,
-                0,
-                0,
-                stepRows.maxLeftInstrinsicWidth + Math.max(
-                    indicators.maxIntrinsicWidth,
-                    lines.maxIntrinsicWidth
-                ),
-                0
-            )
-            val indicatorStartPoint = indicatorMeasureAndPlaceHelpersImpl.firstIndicatorPosition(
-                indicators.firstIndicatorHeight,
-                indicators.firstIndicatorAlignment
-            )
-            indicators.place.invoke(this, indicatorStartPoint.x, indicatorStartPoint.y)
-
-            lines.measure(constraints)
-            val lineStartPoint =
-                lineMeasureAndPlaceHelpersImpl.firstLinePosition(lines.firstLineWidth)
-            lines.place.invoke(this, lineStartPoint.x, lineStartPoint.y)
-        }
-    }
-}
-
-// ---
-
-interface StepRowMeasureAndPlaceHelpers {
-    fun verticalSpacing(belowStepRowWithIndex: Int): Int
-}
-
-class StepRowMeasureAndPlaceHelpersImpl() : StepRowMeasureAndPlaceHelpers {
-    override fun verticalSpacing(belowStepRowWithIndex: Int): Int = VERTICAL_SPACING
-}
-
-interface IndicatorMeasureAndPlaceHelpers {
-    fun verticalSpacing(
-        indicatorHeight: Int,
-        aboveIndicatorWithIndex: Int,
-        alignment: StepIndicatorAlignment
-    ): Int
-
-    fun firstIndicatorPosition(firstIndicatorHeight: Int, alignment: StepIndicatorAlignment): Point
-}
-class Point(val x:Int,val y:Int)
-class IndicatorMeasureAndPlaceHelpersImpl(val stepRows: Rows) : IndicatorMeasureAndPlaceHelpers {
-
-    override fun firstIndicatorPosition(
-        firstIndicatorHeight: Int,
-        alignment: StepIndicatorAlignment
-    ): Point {
-        var yPosition = when (alignment) {
-            StepIndicatorAlignment.CENTER -> stepRows.rows[0].height / 2 - firstIndicatorHeight / 2
-            StepIndicatorAlignment.TOP -> 0
-            StepIndicatorAlignment.BOTTOM -> stepRows.rows[0].height / 2 + firstIndicatorHeight
-        }
-        val xPosition = stepRows.maxLeftInstrinsicWidth
-        return Point(xPosition, yPosition)
-    }
-
-    override fun verticalSpacing(
-        indicatorHeight: Int,
-        aboveIndicatorWithIndex: Int,
-        alignment: StepIndicatorAlignment
-    ): Int {
-        return when (alignment) {
-            StepIndicatorAlignment.TOP -> {
-                if (aboveIndicatorWithIndex == 0)
-                    0
-                else
-                    stepRows.rows[aboveIndicatorWithIndex - 1].height + stepRows.verticalSpacing(
-                        belowStepRowWithIndex = aboveIndicatorWithIndex - 1
-                    )
-            }
-            StepIndicatorAlignment.CENTER -> {
-                if (aboveIndicatorWithIndex == 0)
-                    0
-                else
-                    stepRows.rows[aboveIndicatorWithIndex - 1].height / 2 + stepRows.rows[aboveIndicatorWithIndex].height / 2 + stepRows.verticalSpacing(
-                        belowStepRowWithIndex = aboveIndicatorWithIndex - 1
-                    )
-
-            }
-            StepIndicatorAlignment.BOTTOM -> {
-                if (aboveIndicatorWithIndex == 0)
-                    0
-                else
-                    stepRows.rows[aboveIndicatorWithIndex].height + stepRows.verticalSpacing(
-                        belowStepRowWithIndex = aboveIndicatorWithIndex - 1
-                    )
-            }
-        }
-    }
-}
-
-interface LineMeasureAndPlaceHelpers {
-    fun shouldLineOverlapIndicator(): Boolean
-    fun verticalSpacing(forLineIndex: Int): Int
-    fun lineHeight(forLineIndex: Int): Int
-    fun firstLinePosition(lineWidth: Int): Point
-}
-
-class LineMeasureAndPlaceHelpersImpl(val indicators: StepIndicators) : LineMeasureAndPlaceHelpers {
-    override fun shouldLineOverlapIndicator(): Boolean = false
-
-    override fun verticalSpacing(forLineIndex: Int): Int {
-        return getOverlappingOffset(forLineIndex)
-    }
-
-    override fun lineHeight(forLineIndex: Int): Int {
-        return indicators.components[forLineIndex + 1].yPosition - indicators.components[forLineIndex].yPosition - getOverlappingOffset(
-            forLineIndex
-        )
-    }
-
-    private fun getOverlappingOffset(index: Int): Int {
-        return if (shouldLineOverlapIndicator()) 0 else indicators.getIndicatorHeight(index)
-    }
-
-    override fun firstLinePosition(firstLineWidth: Int): Point {
-        val xPosition =
-            indicators.firstIndicatorPosition.x + (indicators.getIndicatorWidth(0) - firstLineWidth) / 2
-        val yPosition = indicators.firstIndicatorPosition.y + getOverlappingOffset(0)
-        return Point(xPosition, yPosition)
-    }
-}
-
-//---
-@Composable
-fun StepperViewIndicator(modifier: Modifier) {
-    OutlinedButton(
-        onClick = { /*TODO*/ },
-        modifier = modifier.size(10.dp),
-        shape = CircleShape,
-        border = BorderStroke(1.dp, Color.Blue),
-        contentPadding = PaddingValues(0.dp),
-        colors = ButtonDefaults.outlinedButtonColors(contentColor = Color.Blue)
-    ) {
-        //Icon(Icons.Default.Add, contentDescription = "StepperViewIndicator")
-    }
-}
-
-@Composable
-fun StepperViewLine(modifier: Modifier) {
-    Box(
-        modifier = modifier
-            .width(1.dp)
-            .background(Color.Blue)
-    )
-}
-
-enum class StepAlignment {
-    LEFT,
-    RIGHT
-}
-
-class StepAlignmentData(val stepAlignment: StepAlignment) : ParentDataModifier {
-
-    override fun Density.modifyParentData(parentData: Any?): Any? {
-        return this@StepAlignmentData
-    }
-}
-
-interface StepScope {
-    fun Modifier.align(stepAlignment: StepAlignment): Modifier
-}
-
-internal object StepScopeInstance : StepScope {
-
-    override fun Modifier.align(stepAlignment: StepAlignment): Modifier = this.then(
-        StepAlignmentData(stepAlignment)
-    )
-}
-
-enum class StepIndicatorAlignment {
-    TOP,
-    BOTTOM,
-    CENTER
-}
-
-class StepIndicatorAlignmentData(val stepIndicatorAlignment: StepIndicatorAlignment) :
-    ParentDataModifier {
-
-    override fun Density.modifyParentData(parentData: Any?): Any? {
-        return this@StepIndicatorAlignmentData
-    }
-}
-
-interface StepIndicatorScope {
-    fun Modifier.align(stepIndicatorAlignment: StepIndicatorAlignment): Modifier
-}
-
-internal object StepIndicatorScopeInstance : StepIndicatorScope {
-    override fun Modifier.align(stepIndicatorAlignment: StepIndicatorAlignment): Modifier =
-        this.then(
-            StepIndicatorAlignmentData(stepIndicatorAlignment)
-        )
-}
-
-
-// ---- ----
-
-interface StepComponentLayout {
-    val intrinsicWidth: Int
-    val intrinsicHeight: Int
-    var xPosition: Int
-    var yPosition: Int
-    fun measure(constraints: Constraints)
-    var placeable: Placeable?
-    val place: Placeable.PlacementScope.(x: Int, y: Int) -> Unit
-}
-
-class BaseStepComponentLayout(val measurable: Measurable) : StepComponentLayout {
-    override val intrinsicWidth: Int
-        get() = measurable.maxIntrinsicWidth(Int.MAX_VALUE)
-    override val intrinsicHeight: Int
-        get() = measurable.maxIntrinsicHeight(Int.MAX_VALUE)
-    override var placeable: Placeable? = null
-    override var xPosition: Int = 0
-    override var yPosition: Int = 0
-
-    override fun measure(constraints: Constraints) {
-        placeable = measurable.measure(constraints)
-    }
-
-    override val place: Placeable.PlacementScope.(x: Int, y: Int) -> Unit = { x: Int, y: Int ->
-        xPosition = x
-        yPosition = y
-        placeable?.placeRelative(x = x, y = y)
-    }
-}
-
-class Step(val measurable: Measurable) :
-    StepComponentLayout by BaseStepComponentLayout(measurable = measurable) {
-    val stepAlignment
-        get() = measurable.stepAlignment
-}
-
-class StepLine(val measurable: Measurable) :
-    StepComponentLayout by BaseStepComponentLayout(measurable)
-
-class StepIndicator(val measurable: Measurable) :
-    StepComponentLayout by BaseStepComponentLayout(measurable = measurable) {
-    val alignment: StepIndicatorAlignment
-        get() = measurable.stepIndicatorAlignment
-}
-
-class StepIndicators(
-    val indicatorMeasurables: List<Measurable>,
-    indicatorMeasureAndPlaceHelpers: IndicatorMeasureAndPlaceHelpers
-) : BaseComponentList<StepIndicator>(indicatorMeasurables),
-    IndicatorMeasureAndPlaceHelpers by indicatorMeasureAndPlaceHelpers {
-
-    override fun instance(measurable: Measurable) = StepIndicator(measurable)
-
-    fun getIndicatorHeight(index: Int): Int = components[index].placeable?.height ?: 0
-
-    fun getIndicatorWidth(index: Int): Int = components[index].placeable?.width ?: 0
-
-    val firstIndicatorPosition: Point
-        get() = Point(components[0].xPosition, components[0].yPosition)
-    val firstIndicatorHeight
-        get() = getIndicatorHeight(0)
-    val firstIndicatorAlignment
-        get() = components[0].measurable.stepIndicatorAlignment
-
-    fun measure(constraints: Constraints) {
-        components.forEach {
-            it.measure(constraints)
-        }
-    }
-
-    val place: Placeable.PlacementScope.(x: Int, y: Int) -> Unit = { x: Int, y: Int ->
-        var yPosition = y
-        components.forEachIndexed { index, it ->
-            yPosition += indicatorMeasureAndPlaceHelpers.verticalSpacing(
-                it.placeable?.height ?: 0,
-                index,
-                it.alignment
-            )
-            it.place.invoke(this, x, yPosition)
-        }
-    }
-}
-
-abstract class BaseComponentList<T : StepComponentLayout>(measurables: List<Measurable>) {
-    var components = arrayListOf<T>()
-
-    abstract fun instance(measurable: Measurable): T
-
-    init {
-        measurables.forEach {
-            components.add(instance(it))
-        }
-    }
-
-    val maxIntrinsicWidth : Int
-        get(){
-            return components.maxOfOrNull { it.intrinsicWidth } ?: 0
-        }
-}
-
-
-class StepLines(
-    lineMeasurables: List<Measurable>,
-    private val lineMeasureAndPlaceHelpers: LineMeasureAndPlaceHelpers
-) : BaseComponentList<StepLine>(lineMeasurables) {
-
-    override fun instance(measurable: Measurable) = StepLine(measurable)
-
-    val firstLineWidth
-        get() = components[0].placeable?.width ?: 0
-
-    fun measure(constraints: Constraints) {
-        components.forEachIndexed { index, it ->
-            it.measure(
-                constraints = constraints.copy(
-                    minHeight = lineMeasureAndPlaceHelpers.lineHeight(index),
-                    maxHeight = lineMeasureAndPlaceHelpers.lineHeight(index)
-                )
-            )
-        }
-    }
-
-    val place: Placeable.PlacementScope.(x: Int, y: Int) -> Unit = { x, y ->
-        var yPosition = y
-        components.forEachIndexed { index, it ->
-            it.place.invoke(this, x, yPosition)
-            yPosition += (it.placeable?.height ?: 0) + lineMeasureAndPlaceHelpers.verticalSpacing(
-                index
-            )
-        }
-    }
-}
-
-class StepRow(stepMeasurables: List<Measurable>) {
-
-    var steps = arrayListOf<Step>()
-
-    init {
-        stepMeasurables.forEach { steps.add(Step(it)) }
-    }
-
-    var height = 0
-        get() = steps.maxOfOrNull { it.placeable?.height ?: 0 } ?: 0
-
-//    val maxIntrinsicWidth = steps.maxOfOrNull { it.intrinsicWidth } ?: 0
-//    val maxIntrinsicHeight = steps.maxOfOrNull { it.intrinsicHeight } ?: 0
-
-    fun measure(leftConstraints: Constraints, rightConstraints: Constraints) {
-        steps.forEach {
-            when (it.stepAlignment) {
-                StepAlignment.LEFT -> it.measure(leftConstraints)
-                StepAlignment.RIGHT -> it.measure(rightConstraints)
-            }
-        }
-    }
-
-    val place: Placeable.PlacementScope.(x: Int, y: Int, x1: Int, y1: Int) -> Unit =
-        { x: Int, y: Int, x1: Int, y1: Int ->
-            steps.forEach {
-                if (it.stepAlignment == StepAlignment.LEFT) {
-                    it.place.invoke(this, x, y)
-                } else {
-                    it.place.invoke(this, x1, y1)
-                }
-            }
-        }
-}
-
-class Rows(
-    stepMeasurables: List<Measurable>,
-    val stepsPerRow: StepsPerRow,
-    stepRowMeasureAndPlaceHelpers: StepRowMeasureAndPlaceHelpers
-) : StepRowMeasureAndPlaceHelpers by stepRowMeasureAndPlaceHelpers {
-
-    val rows = arrayListOf<StepRow>()
-
-    init {
-        for (i in stepMeasurables.indices step stepsPerRow.noOfSteps) {
-            rows.add(
-                StepRow(
-                    stepMeasurables.toMutableList().subList(i, i + stepsPerRow.noOfSteps)
-                        .toList())
-            )
-        }
-    }
-
-    val steps
-        get() = rows.flatMap { it.steps }
-    val leftSteps
-        get() = steps.filter { it.stepAlignment == StepAlignment.LEFT }
-    val rightSteps
-        get() = steps.filter { it.stepAlignment == StepAlignment.RIGHT }
-    val maxLeftInstrinsicWidth
-        get() = leftSteps.maxOfOrNull { it.intrinsicWidth } ?: 0
-
-    val maxRightInstrinsicWidth
-        get() = rightSteps.maxOfOrNull { it.intrinsicWidth } ?: 0
-
-    fun remainingLeftStepWidth(constraints: Constraints): Int {
-        val maxWidthEqualConstraints = equalConstraints(constraints = constraints).maxWidth
-        if (maxLeftInstrinsicWidth >= maxWidthEqualConstraints) return 0
-        else return maxWidthEqualConstraints - maxLeftInstrinsicWidth
-    }
-
-    fun remainingRightStepWidth(constraints: Constraints): Int {
-        val maxWidthEqualConstraints = equalConstraints(constraints = constraints).maxWidth
-        if (maxRightInstrinsicWidth >= maxWidthEqualConstraints) return 0
-        else return maxWidthEqualConstraints - maxRightInstrinsicWidth
-    }
-
-    var isBiDirectionalStepper: Boolean = false
-        get() {
-            val steps = rows.flatMap { it.steps }
-            val leftSteps = steps.filter { it.stepAlignment == StepAlignment.LEFT }
-            val rightSteps = steps.filter { it.stepAlignment == StepAlignment.RIGHT }
-            return (leftSteps.isNotEmpty() && rightSteps.isNotEmpty())
-        }
-
-    fun measure(constraints: Constraints) {
-        if (!isBiDirectionalStepper) {
-            rows.forEach { it.measure(constraints, constraints) }
-        } else {
-            rows.forEach { it.measure(leftConstraints(constraints), rightConstraints(constraints)) }
-        }
-    }
-
-    val place: Placeable.PlacementScope.(x: Int, y: Int, x1: Int, y1: Int) -> Unit =
-        { x: Int, y: Int, x1: Int, y1: Int ->
-            var yPosition = Math.max(y, y1)
-            rows.forEachIndexed { index, stepRow ->
-                stepRow.place.invoke(this, x, yPosition, x1, yPosition)
-                yPosition += stepRow.height + stepRowMeasureAndPlaceHelpers.verticalSpacing(
-                    belowStepRowWithIndex = index
-                )
-            }
-        }
-
-    private fun equalConstraints(constraints: Constraints) = constraints.copy(
-        maxWidth = constraints.maxWidth / 2
-    )
-
-    private fun leftConstraints(constraints: Constraints): Constraints {
-        if (maxLeftInstrinsicWidth >= equalConstraints(constraints = constraints).maxWidth)
-            return constraints
-        else
-            return constraints.copy(
-                maxWidth = equalConstraints(constraints = constraints).maxWidth + remainingRightStepWidth(
-                    constraints
-                )
-            )
-    }
-
-    private fun rightConstraints(constraints: Constraints): Constraints {
-        if (maxRightInstrinsicWidth >= equalConstraints(constraints = constraints).maxWidth)
-            return constraints
-        else
-            return constraints.copy(
-                maxWidth = equalConstraints(constraints = constraints).maxWidth + remainingLeftStepWidth(
-                    constraints
-                )
-            )
-        return constraints
-    }
-
-}
-
