+++ stepperview/src/main/java/com/ss/stepperview/data/Point.kt
@@ -1,0 +1,3 @@
+package com.ss.stepperview.data
+
+class Point(val x:Int,val y:Int)
