+++ stepperview/src/main/java/com/ss/stepperview/layout/helpers/StepRowMeasureAndPlaceHelpers.kt
@@ -1,0 +1,11 @@
+package com.ss.stepperview.layout.helpers
+
+private const val VERTICAL_SPACING = 20
+
+interface StepRowMeasureAndPlaceHelpers {
+    fun verticalSpacing(belowStepRowWithIndex: Int): Int
+}
+
+class StepRowMeasureAndPlaceHelpersImpl : StepRowMeasureAndPlaceHelpers {
+    override fun verticalSpacing(belowStepRowWithIndex: Int): Int = VERTICAL_SPACING
+}
