+++ stepperview/src/main/java/com/ss/stepperview/layoutmodifier/StepIndicatorAlignment.kt
@@ -1,0 +1,34 @@
+package com.ss.stepperview.layoutmodifier
+
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.Measurable
+import androidx.compose.ui.layout.ParentDataModifier
+import androidx.compose.ui.unit.Density
+
+enum class StepIndicatorAlignment {
+    TOP,
+    BOTTOM,
+    CENTER
+}
+
+internal class StepIndicatorAlignmentData(val stepIndicatorAlignment: StepIndicatorAlignment) :
+    ParentDataModifier {
+    override fun Density.modifyParentData(parentData: Any?): Any {
+        return this@StepIndicatorAlignmentData
+    }
+}
+
+internal interface StepIndicatorScope {
+    fun Modifier.align(stepIndicatorAlignment: StepIndicatorAlignment): Modifier
+}
+
+internal object StepIndicatorScopeInstance : StepIndicatorScope {
+    override fun Modifier.align(stepIndicatorAlignment: StepIndicatorAlignment): Modifier =
+        this.then(
+            StepIndicatorAlignmentData(stepIndicatorAlignment)
+        )
+}
+
+val Measurable.stepIndicatorAlignment
+    get() = (parentData as? StepIndicatorAlignmentData)?.stepIndicatorAlignment
+        ?: StepIndicatorAlignment.CENTER
