+++ stepperview/src/main/java/com/ss/stepperview/view/StepperViewIndicator.kt
@@ -1,0 +1,26 @@
+package com.ss.stepperview.view
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material.ButtonDefaults
+import androidx.compose.material.OutlinedButton
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun StepperViewIndicator(modifier: Modifier) {
+    OutlinedButton(
+        onClick = { /*TODO*/ },
+        modifier = modifier.size(10.dp),
+        shape = CircleShape,
+        border = BorderStroke(1.dp, Color.Blue),
+        contentPadding = PaddingValues(0.dp),
+        colors = ButtonDefaults.outlinedButtonColors(contentColor = Color.Blue)
+    ) {
+        //Icon(Icons.Default.Add, contentDescription = "StepperViewIndicator")
+    }
+}
