+++ stepperview/src/main/java/com/ss/stepperview/view/StepperViewLine.kt
@@ -1,0 +1,18 @@
+package com.ss.stepperview.view
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.width
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun StepperViewLine(modifier: Modifier) {
+    Box(
+        modifier = modifier
+            .width(1.dp)
+            .background(Color.Blue)
+    )
+}
