+++ stepperview/src/main/java/com/ss/stepperview/view/StepperViewStep.kt
@@ -1,0 +1,10 @@
+package com.ss.stepperview.view
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun Step(
+    content: @Composable () -> Unit
+) {
+    content()
+}
